+++ Shoko.Server/Providers/AniDB/Interfaces/IConnectionHandler.cs
@@ -5,7 +5,7 @@ namespace Shoko.Server.Providers.AniDB.Interfaces
     public interface IConnectionHandler
     {
         string Type { get; }
-        int BanTimerResetLength { get; }
+        double BanTimerResetLength { get; }
         DateTime? BanTime { get; set; }
         event EventHandler<AniDBStateUpdate> AniDBStateUpdate;
         AniDBStateUpdate State { get; set; }
